requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
pypdf
faiss-cpu
duckduckgo-search
python-dotenv       